backend/requirements.txt
Flask==2.3.3
Flask-CORS==3.0.10
beautifulsoup4==4.12.2
requests==2.30.0
python-dotenv==1.0.0
